requirements.txt
tk
dnspython
python-whois
